refactor(web): split useOnlineSkipBoGame into focused modules
The online hook had grown to 1070 lines mixing six concerns. Decompose it
into single-responsibility modules under hooks/useOnlineSkipBoGame/ with no
behavior change (the host hook still owns all shared state, refs, host
orchestration, and the interaction handlers):

- useOnlineConnection.ts — the WebSocket lifecycle (connect / auth / ping
  keepalive / backoff reconnect) and ServerMessage routing. Owns the socket,
  ping, and reconnect refs internally; reads its collaborators through a
  synced ref so a re-render never tears down a live socket. Re-runs only on
  session change, exactly as before.
- localActionAnimations.ts — startPlayCardAnimation / startDiscardCardAnimation,
  the DOM-measured play/discard animation builders lifted verbatim out of
  playCard / discardCard.
- types.ts — ConnectionStatus, HostSnapshotPayload, toRoomMeta.

The host hook drops from 1070 to 607 lines and reads as a clear composition.

Verified behavior-preserving by the full suite: typecheck, lint, all 254
unit tests (including the 1377-line mock-WebSocket online-hook suite that
exercises reconnect, message routing, lobby flow, optimistic updates, and
opponent/own-move animations), and a production build.
